pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "equity-valuation-engine"
version = "0.1.0"
description = "A valuation engine for equities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yfinance>=1.2.2",
    "pandas>=3.0.2",
    "requests>=2.33.1",
    "PyYAML>=6.0.3",
    "tabulate >= 0.10.0"
]

[tool.setuptools.packages.find]
where = ["src"]